requirements.txt
pyzmq
rocksdict
msgpack
cryptography